airflow_home/requirements.txt
pyspark==3.5.2
kafka-python
pandas
mysql-connector-python
SQLAlchemy
apache-airflow==2.10.1
apache-airflow-providers-mysql==5.7.0
Flask-Session==0.5.0
apache-airflow-providers-apache-spark==4.10.0